+++ src/dom/attribute.spec.ts
@@ -1,10 +1,10 @@
 import { Attribute } from './attribute';
-import { Location } from 'context';
+import { Location } from '../context';
 
 describe('Attribute', () => {
 
 	it('should set fields', () => {
-		const location: Location = {filename: 'file', line: 1, column: 2};
+		const location: Location = {filename: 'file', offset: 1, line: 1, column: 2};
 		const attr = new Attribute('foo', 'bar', location);
 		expect(attr.key).toEqual('foo');
 		expect(attr.value).toEqual('bar');