+++ src/reporter.spec.ts
@@ -64,11 +64,11 @@ describe('Reporter', function(){
 			report.addManual("bar.html", createMessage("another error", 2));
 			expect(report.save().results).toEqual([
 				{filePath: 'foo.html', errorCount: 1, warningCount: 1, messages: [
-					{line: 1, column: 1, ruleId: 'mock', severity: 2, message: 'error'},
-					{line: 1, column: 1, ruleId: 'mock', severity: 1, message: 'warning'},
+					{offset: 0, line: 1, column: 1, ruleId: 'mock', severity: 2, message: 'error'},
+					{offset: 0, line: 1, column: 1, ruleId: 'mock', severity: 1, message: 'warning'},
 				]},
 				{filePath: 'bar.html', errorCount: 1, warningCount: 0, messages: [
-					{line: 1, column: 1, ruleId: 'mock', severity: 2, message: 'another error'},
+					{offset: 0, line: 1, column: 1, ruleId: 'mock', severity: 2, message: 'another error'},
 				]},
 			]);
 		});
@@ -89,6 +89,7 @@ function createMessage(message: string, severity: number = 2): Message {
 		ruleId: 'mock',
 		severity,
 		message,
+		offset: 0,
 		line: 1,
 		column: 1,
 	};